Revert "merge everything into a single file"
This reverts commit f96c56f8bd483cad3f4ad2340cfa4959b9004b50.
