+++ .gitignore
@@ -1,6 +1,4 @@
 .idea
 *.iml
 __pycache__/
-**/*.pyc
 *.pyc
-venv
