--- app.cfg
@@ -1,4 +1,0 @@
-SQLALCHEMY_DATABASE_URI = 'postgres:///cphack'
-SQLALCHEMY_ECHO = False
-SECRET_KEY = '\xfb\x12\xdf\xa1@i\xd6>V\xc0\xbb\x8fp\x16#Z\x0b\x81\xeb\x16'
-DEBUG = True
